binary cipher
a program for encrypting a message using a binary sequence created using recurrent chaotic mapping:
xn+1 = 3.9999xn(1 − xn), 0 < xn < 1, n = 0, 1, . . .
Data for encryption is a text message or a file of any format. To obtain a binary sequence for encryption, the initial is set
condition of chaotic mapping, for example, x0 = 0.1234567. Next, a sequence of the required length equal to the number of bits in the message or file is generated recursively using chaotic mapping. The obtained sequence
of real numbers is converted into a binary sequence
The encrypted text is obtained by bitwise addition modulo 2 (XOR)
message/file bn and bit sequences.
The encrypted text is obtained by bitwise addition modulo 2 (XOR)
message/file bn and bit sequences.
During decryption, a chaotic sequence is generated with the same initial condition as during encryption. To get the decrypted text/file
it is necessary to bitwise add modulo 2 (XOR) the sequence bn and the bits of the encrypted message/file (ciphertext).