matrix game
Find the differences
a program that simulates the game "Find the differences" between matrices
two integers. The program for a short time, for example 3 seconds, displays on
the screen is two matrices of integers that differ by several values ​​depending on
from the level of complexity. The player's task is to find differences. After 10
of work cycles, the program counts the number of correct answers and reports
players
Main characteristics: (Main menu of the program)
• New game
• Show the best results
• Settings
• Output
Game settings submenu. The submenu opens when you select the appropriate menu item. Select the size of the playing field (by default 3*3).
• Field size 3*3, 1 mismatch
• Field size 4*4, 1 mismatch
• Field size 4*4, 2 mismatches
• Field size 5*5, 1 mismatch
• Field size 5*5, 2 mismatches
• Field size 5*5, 3 mismatches